backend/requirements.txt
wheel
setuptools
pillow==10.3.0
flask==3.0.3
flask-caching==2.1.0
elasticsearch==8.13.0
pydantic==2.7.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
redis==5.0.3
async_timeout==4.0.3
SQLAlchemy==2.0.29
gunicorn==23.0.0
tqdm==4.66.3
requests==2.32.0
tiktoken==0.6.0
openai==1.66.3
spacy==3.7.4
func-timeout==4.3.5
markdown==3.6
Flask-Limiter[redis]==3.6.0
flask_sqlalchemy==3.1.1
mysqlclient==2.2.4
elasticsearch-dsl==8.15.4
beautifulsoup4==4.12.3
sentence-transformers==3.4.1
bleach==6.2.0
meilisearch==0.34.1
geoip2==5.1.0